refactor: extract shared file validation utility and move interface to module level
- Move FirstRunAttachment interface from component body to module level (P1)
- Extract validateFiles() to shared lib/validate-files.ts (P2)
- Both dashboard/page.tsx and chat-input.tsx now use shared validation
